requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow>=9.0.0
python-decouple==3.8
django-extensions==3.2.3
